requirements.txt
Django==5.0.3
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
pytest==8.0.2
pytest-django==4.8.0
pytest-cov==4.0.0
ruff==0.3.0
black==24.2.0
coverage==7.4.3